+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Warn if a string is passed directly as a tags parameter, which might cause unexpected behavior.
 
 ## [3.1.0] - 2023-06-10
 ### Changed