+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [3.2.0] - 2023-07-02
 ### Changed
 - Warn if a string is passed directly as a tags parameter, which might cause unexpected behavior.
 - `Entity.relation_tags` has been renamed to `Entity.relation_tag`.